Move metrics hooks to own file + re-use useStatementLookup
